http://jira.aplana.com/browse/SBRFACCTAX-12116 0.6.1 Неверный расчет по "длинным цифрам" найден в декларации НДС. Просмотреть скрипты всех НФ и деклараций для поиска такой же проблемы.

+++ src/main/resources/form_template/income/advanceDistribution/v2015/script.groovy
@@ -231,7 +231,7 @@ void calc() {
             def departmentParam = getRefBookValue(30, row.regionBankDivision)
 
             def depParam = getDepParam(departmentParam)
-            def depId = depParam.get(RefBook.RECORD_ID_ALIAS).numberValue as int
+            def depId = depParam.get(RefBook.RECORD_ID_ALIAS).numberValue as long
 
             incomeParam = getRefBookRecord(33, "DEPARTMENT_ID", "$depId", departmentParamsDate, -1, null, false)
         }
@@ -526,7 +526,7 @@ void logicalCheckBeforeCalc() {
 
         // Определение условий для проверок 2, 3, 4
         def depParam = getDepParam(departmentParam)
-        def depId = depParam.get(RefBook.RECORD_ID_ALIAS).numberValue as int ?: -1
+        def depId = depParam.get(RefBook.RECORD_ID_ALIAS).numberValue as long ?: -1
         def departmentName = depParam?.NAME?.stringValue ?: "Не задано"
         def incomeParam = getProvider(33).getRecords(getReportPeriodEndDate() - 1, null, "DEPARTMENT_ID = $depId", null)
         def incomeParamTable = getIncomeParamTable(depParam)
@@ -1380,7 +1380,7 @@ def getDepParam(def departmentParam) {
 
 // Получение параметров (справочник 330)
 def getIncomeParamTable(def depParam) {
-    def depId = depParam.get(RefBook.RECORD_ID_ALIAS).numberValue as int
+    def depId = depParam.get(RefBook.RECORD_ID_ALIAS).numberValue as long
     def incomeParam = getProvider(33).getRecords(getReportPeriodEndDate() - 1, null, "DEPARTMENT_ID = $depId", null)
     if (incomeParam != null && !incomeParam.isEmpty()) {
         def link = incomeParam.get(0).record_id.value
